+++ Changelog.rst
@@ -16,6 +16,10 @@ OfflineIMAP v6.5.6 (YYYY-MM-DD)
   adds mechanics to change message labels (Abdó Roig-Maranges)
 * Allow to migrate status data across differend backends
   (Abdó Roig-Maranges)
+* Support XDG Base Directory Specification
+  (if $XDG_CONFIG_HOME/offlineimap/config exists, use it as the
+  default configuration path; ~/.offlineimaprc is still tried after
+  XDG location) (GitHub#32)
 
 
 OfflineIMAP v6.5.5 (2013-10-07)